+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 === HEAD
 
+* Add per-file SUIT conformance checking.
+* Remove all preprocessor functionality.
+
 === 0.1.0 (February 8, 2014)
 
-Initial release.
+* Initial release.